Create RegisterNotExistsException. Adjust ProductServiceTest structure. Create delete product tests

+++ WPFSample/WPFSample.Service/Implementation/ProductService.cs
@@ -138,6 +138,9 @@ namespace WPFSample.Service.Implementation
         {
             Product product = _productRepository.GetById(id);
 
+            if (product == null)
+                throw new RegisterNotExistsException("Register non-existing");
+
             _productRepository.Delete(product);
         }
